Move Arrow Core test to commonTest
