+++ arrow-libs/core/arrow-continuations/src/commonTest/kotlin/generic/MaybeEffect.kt
