+++ arrow-libs/core/arrow-core-test/src/commonMain/kotlin/arrow/core/test/laws/Law.kt
