+++ arrow-libs/core/arrow-core-test/src/commonMain/kotlin/arrow/core/test/laws/SemiringLaws.kt
