+++ arrow-libs/core/arrow-core-test/src/jvmMain/kotlin/arrow/core/test/generators/GeneratorsJvm.kt
@@ -1,0 +1,12 @@
+package arrow.core.test.generators
+
+import arrow.core.Either
+import io.kotest.property.Arb
+import io.kotest.property.arbitrary.map
+import io.kotest.property.arbitrary.of
+
+fun Arb.Companion.suspendFunThatThrowsFatalThrowable(): Arb<suspend () -> Either<Any, Any>> =
+  fatalThrowable().map { suspend { throw it } } as Arb<suspend () -> Either<Any, Any>>
+
+fun Arb.Companion.fatalThrowable(): Arb<Throwable> =
+  Arb.of(listOf(ThreadDeath(), StackOverflowError(), OutOfMemoryError(), InterruptedException()))
