--- arrow-libs/core/arrow-core-test/src/main/kotlin/arrow/Kind.kt
@@ -1,6 +1,0 @@
-package arrow
-
-/**
- * TODO Used only for tests runtime until Kotlintest is upgraded to Kotest
- */
-interface Kind<F, out A>
