+++ arrow-libs/core/arrow-core/build.gradle
@@ -20,6 +20,11 @@ kotlin {
                 api project(":arrow-annotations")
             }
         }
+        commonTest {
+            dependencies {
+                implementation project(":arrow-core-test")
+            }
+        }
         jvmMain {
             dependencies {
                 compileOnly "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$KOTLIN_VERSION"
@@ -28,9 +33,12 @@ kotlin {
 
         jvmTest {
             dependencies {
-                compileOnly project(":arrow-core-test")
-                runtimeOnly "org.junit.vintage:junit-vintage-engine:$JUNIT_VINTAGE_VERSION"
+                runtimeOnly "io.kotest:kotest-runner-junit5:$KOTEST_VERSION"
             }
         }
     }
+}
+
+test {
+    useJUnitPlatform()
 }
