+++ arrow-libs/core/arrow-core/src/jvmTest/java/arrow/core/ValidatedUsage.java
